Initial roster push

+++ api/users.js
@@ -2,7 +2,8 @@
 const router = require('express').Router();
 const validation = require('../lib/validation');
 const { generateAuthToken,
-      requireAuthentication } = require('../lib/auth');
+      requireAuthentication,
+      requireAuthorizationURL} = require('../lib/auth');
 
 
 const {
@@ -20,7 +21,7 @@ const userSchema = {
     role: { required: true }
 }
 
-router.get('/', async (req, res, next) => {
+router.get('/', requireAuthentication, async (req, res, next) => {
    try{
        //if authenticated
         //send user
@@ -34,13 +35,19 @@ router.get('/', async (req, res, next) => {
    }
 });
 
-router.get('/:id', async (req, res, next) => {
+router.get('/:id', requireAuthentication, async (req, res, next) => {
   try {
-    const user = await getUserDetailsbyID(req.params.id);
-    if (user) {
-      res.status(200).send(user);
+    if(requireAuthorizationURL(req, res, next, 'id') == 1){
+      const user = await getUserDetailsbyID(req.params.id);
+      if (user) {
+        res.status(200).send(user);
+      } else {
+        next();
+      }
     } else {
-      next();
+      res.status(401).send({
+        error: "User is not authorized to perform this action"
+      });
     }
   } catch (err) {
     console.error(err);
@@ -88,6 +95,8 @@ router.post('/login', async (req, res) => {
 
 router.post('/', async (req, res, next) => {
    try {
+       console.log("Body: ", req.body);
+       console.log("Validation: ", validation.validateAgainstSchema(req.body, userSchema));
        if(validation.validateAgainstSchema(req.body, userSchema)){
         //Insert into mongoDB
         const id = await insertNewUser(req.body);
@@ -117,23 +126,29 @@ router.post('/', async (req, res, next) => {
    }
 });
 
-router.put('/:id', async (req, res, next) => {
+router.put('/:id', requireAuthentication,async (req, res, next) => {
    try{
-       if(validation.validateAgainstSchema(req.body, userSchema)){
-        //Insert into mongoDB
-        const result = await updateUser(req.body, req.params.id);
-        res.status(201).send({
-            res: result,
-            id: req.params.id,
-            links: {
-                user: `/users/${req.params.id}`
-            }
-         });
-       } else {
-         res.status(400).send({
-             error: "invalid body"
-         });
-       }
+      if(requireAuthorizationURL(req, res, next, 'id') == 1){  
+         if(validation.validateAgainstSchema(req.body, userSchema)){
+          //Insert into mongoDB
+          const result = await updateUser(req.body, req.params.id);
+          res.status(201).send({
+              res: result,
+              id: req.params.id,
+              links: {
+                  user: `/users/${req.params.id}`
+              }
+           });
+         } else {
+           res.status(400).send({
+               error: "invalid body"
+           });
+         }
+      } else {
+        res.status(401).send({
+          error: "User is not authorized to perform this action"
+        });
+      }
        
    } catch (err) {
        console.error(err);
@@ -143,10 +158,16 @@ router.put('/:id', async (req, res, next) => {
    }
 });
 
-router.delete('/:id', async (req, res, next) => {
+router.delete('/:id', requireAuthentication, async (req, res, next) => {
    try{
-     var result = await deleteUser(req.params.id);
-     res.status(201).send({status: "deleted", result: result});
+     if(requireAuthorizationURL(req, res, next, 'id') == 1){  
+       var result = await deleteUser(req.params.id);
+       res.status(201).send({status: "deleted", result: result});
+     } else {
+       res.status(401).send({
+          error: "User is not authorized to perform this action"
+        });
+     }
    } catch (err){
        console.error(err);
        res.status(500).send({
